# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API TJJud",
        "description": "Documentação da API do sistema de gerenciamento de livros, autores e assuntos.",
        "contact": {
            "name": "Seu Nome",
            "email": "seu-email@exemplo.com"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
        "/api/assuntos": {
            "get": {
                "tags": [
                    "Assuntos"
                ],
                "summary": "Listar assuntos",
                "operationId": "d508d3604432e32082c9f627ffd78c66",
                "responses": {
                    "200": {
                        "description": "Lista de assuntos"
                    }
                }
            },
            "post": {
                "tags": [
                    "Assuntos"
                ],
                "summary": "Criar assunto",
                "operationId": "af3c6b92c90b765ccc4e3cd5dfb28ecd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "Descricao"
                                ],
                                "properties": {
                                    "Descricao": {
                                        "type": "string",
                                        "example": "Assunto Teste"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Assunto criado com sucesso"
                    },
                    "500": {
                        "description": "Erro ao criar assunto"
                    }
                }
            }
        },
        "/api/assuntos/{codAs}": {
            "get": {
                "tags": [
                    "Assuntos"
                ],
                "summary": "Exibir assunto",
                "operationId": "53833795dd115c7c51305d95774fd5b0",
                "parameters": [
                    {
                        "name": "codAs",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Assunto encontrado"
                    },
                    "404": {
                        "description": "Assunto não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Assuntos"
                ],
                "summary": "Atualizar assunto",
                "operationId": "b6479aca543f801a81e75daf146d8719",
                "parameters": [
                    {
                        "name": "codAs",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "Descricao"
                                ],
                                "properties": {
                                    "Descricao": {
                                        "type": "string",
                                        "example": "Assunto Atualizado"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Assunto atualizado com sucesso"
                    },
                    "404": {
                        "description": "Assunto não encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Assuntos"
                ],
                "summary": "Excluir assunto",
                "operationId": "d6290df314a0e8110935b2672422ab16",
                "parameters": [
                    {
                        "name": "codAs",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Assunto excluído com sucesso"
                    },
                    "404": {
                        "description": "Assunto não encontrado"
                    }
                }
            }
        },
        "/api/autores": {
            "get": {
                "tags": [
                    "Autores"
                ],
                "summary": "Listar autores",
                "operationId": "347c686fedd9f10a6e78a12949be3ca9",
                "responses": {
                    "200": {
                        "description": "Lista de autores"
                    }
                }
            },
            "post": {
                "tags": [
                    "Autores"
                ],
                "summary": "Criar autor",
                "operationId": "bc9bd4b13f84f2b5568af1b097c49d48",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "Nome"
                                ],
                                "properties": {
                                    "Nome": {
                                        "type": "string",
                                        "example": "Autor Teste"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Autor criado com sucesso"
                    },
                    "500": {
                        "description": "Erro ao criar autor"
                    }
                }
            }
        },
        "/api/autores/{codAu}": {
            "get": {
                "tags": [
                    "Autores"
                ],
                "summary": "Exibir autor",
                "operationId": "51e4f6c89c5e2450da5b492a6c812054",
                "parameters": [
                    {
                        "name": "codAu",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Autor encontrado"
                    },
                    "404": {
                        "description": "Autor não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Autores"
                ],
                "summary": "Atualizar autor",
                "operationId": "d70bb599e9e99bc9cec33cb201d0a2e8",
                "parameters": [
                    {
                        "name": "codAu",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "Nome"
                                ],
                                "properties": {
                                    "Nome": {
                                        "type": "string",
                                        "example": "Autor Atualizado"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Autor atualizado com sucesso"
                    },
                    "404": {
                        "description": "Autor não encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Autores"
                ],
                "summary": "Excluir autor",
                "operationId": "f5ca8fb5978e5660e33250092f94d7d7",
                "parameters": [
                    {
                        "name": "codAu",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Autor excluído com sucesso"
                    },
                    "404": {
                        "description": "Autor não encontrado"
                    }
                }
            }
        },
...
    },
...
}